fixes dropping of spans without decoration
