+++ mermin/src/k8s.rs
@@ -1,7 +1,7 @@
+pub mod attributor;
 pub mod decorator;
 pub mod error;
 pub mod owner_relations;
-pub mod parser;
 pub mod selector_relations;
 
 pub use error::K8sError;
